requirements.txt
asgiref==3.2.7
boto3==1.13.6
botocore==1.16.6
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-environ==0.4.5
django-extensions==2.2.9
django-model-utils==4.0.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-yasg==1.17.1
gevent==20.5.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.3
Pillow==7.1.2
pyasn1==0.4.8
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
rsa==3.4.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
supervisor==4.2.0
uritemplate==3.0.1
urllib3==1.25.9
sentry-sdk==0.16.1
PyMySQL==0.10.1
django-debug-toolbar==3.1.1